List editor p2 (#7)
Add to listEditorClasses
-UpdateLegoList in Contr
-Error Checks to remove add and update methods
-created tests for all current methods (excluding view methods)
Created Display of all legos in db
-on home page and listEditor page
-formatted in CSS
-created Home classes, shared Lego class, and added to listEditor
Fixed Errors
-List name Regex with "'"
-Deleted stats index.php

* add updateLegoList to listEditorContrClass, no test made

* add error checks to the remove add and update methodes

* Added fetures to listEditor, created listEditor unit tests, added/formated display of legos in db on home and listEditor, errors fixed

+++ src/user/listEditor/ListEditorContrClass.php
@@ -4,15 +4,20 @@ declare(strict_types = 1);
 namespace Src\User\ListEditor;
 require __DIR__ . '\\..\\..\\..\\vendor\\autoload.php';
 
-use Src\Shared\Exceptions\StmtFailedException;
-use Src\User\Add\Lego\LegoClass;
+use Src\Shared\Exceptions\InvalidInputException;
 use Src\User\ListEditor\ListEditorClass;
 use Src\User\ListEditor\ListEditorViewClass;
+use Src\User\Add\Lego\LegoClass;
+use Src\User\Add\LegoList\LegoListContrClass;
+use Src\Shared\Classes\LegoClass as LegoClassS;
 
-class ListEditorContrClass {
+class ListEditorContrClass extends LegoListContrClass {
 
 	private $listEditorClass;
 	private $listEditorViewClass;
+	private $legoClass;
+	private $legoClassS;
+
 	private ?string $eMessage = null;
 	private array $legoListVals = array(
 		'listId'=>null,
@@ -24,12 +29,19 @@ class ListEditorContrClass {
 	);
 	private array $legoListLegos = array();
 
-	public function __construct($legoClass = null, $listEditorClass = null, $listEditorViewClass = null) {
+	public function __construct($legoClass = null, $legoClassS = null, $listEditorClass = null, $listEditorViewClass = null) {
 		if (is_null($legoClass)) {
 			$this->legoClass = new LegoClass();
 		}
 		else {
 			$this->legoClass = $legoClass;
+		}
+
+		if (is_null($legoClassS)) {
+			$this->legoClassS = new LegoClassS();
+		}
+		else {
+			$this->legoClassS = $legoClassS;
 		}
 
 		if (is_null($listEditorClass)) {
@@ -39,12 +51,38 @@ class ListEditorContrClass {
 			$this->listEditorClass = $listEditorClass;
 		}
 
-		if (is_null($listEditorClass)) {
+		if (is_null($listEditorViewClass)) {
 			$this->listEditorViewClass = new ListEditorViewClass();
 		}
 		else {
 			$this->listEditorViewClass = $listEditorViewClass;
 		}
+	}
+
+	public function setLegoListVals(array $legoListVals): void
+	{
+		foreach ($legoListVals as $key => $value) {
+			$this->legoListVals[$key] = $value;
+		}
+	}
+
+	public function getLegoListVals(bool $getFull = true, array $valNames = array()): array
+	{
+		if ($getFull) {
+			return $this->legoListVals;
+		}
+		else {
+			$returnArray = array();
+			foreach ($valNames as $key) {
+				$returnArray[$key] = $this->legoListVals[$key];
+			}
+			return $returnArray;
+		}
+	}
+
+	public function getLegoListLegos(): array
+	{
+		return $this->legoListLegos;
 	}
 
 
@@ -56,7 +94,7 @@ class ListEditorContrClass {
 		return $checkResults[0];
 	}
 
-	public function getEMessage(): string {
+	public function getEMessage(): ?string {
 		return $this->eMessage;
 	}
 
@@ -77,18 +115,80 @@ class ListEditorContrClass {
 
 	// post data manipulater methods
 	public function updateLegoList(array $legoListVals): void {
+		// run error checks
+		// check if listId is valid and if current user has permision to edit
+		if (!$this->checkListId($legoListVals['listId'])) {
+			// echo 'Invalid ListId';
+			throw new InvalidInputException('listid: ' . $this->getEMessage());
+		}
 
+		// create new array with reformated data
+		$validLegoListVals = $this->parentChecks($legoListVals);
+		$validLegoListVals['listId'] = $legoListVals['listId'];
+
+		// update the list
+		$this->listEditorClass->updateLegoListData($validLegoListVals);
 	}
 
 	public function addLegoToLegoList(array $addLegoVals): void {
+		// run error checks
+		// check if listId is valid and if current user has permision to edit
+		if (!$this->checkListId($addLegoVals['listId'])) {
+			// echo 'Invalid ListId';
+			throw new InvalidInputException('listid: ' . $this->getEMessage());
+		}
+
+		// check if lego exists
+		if (!$this->legoClass->checkLegoExist($addLegoVals['legoId'])) {
+			// echo 'Lego does not Exists';
+			throw new InvalidInputException('legonotexists');
+		}
+
+		// check if lego is not in list
+		if ($this->listEditorClass->checkLegoInLegoList($addLegoVals['listId'], $addLegoVals['legoId'])) {
+			// echo 'Lego already in list';
+			throw new InvalidInputException('legoinlist');
+		}
+
 		$this->listEditorClass->setLegoToList($addLegoVals);
 	}
 
 	public function removeLegoFromLegoList(array $removeLegoVals): void {
+		// run error checks
+		// check if listId is valid and if current user has permision to edit
+		if (!$this->checkListId($removeLegoVals['listId'])) {
+			// echo 'Invalid ListId';
+			throw new InvalidInputException('listid: ' . $this->getEMessage());
+		}
+
+		// check if lego exists
+		if (!$this->legoClass->checkLegoExist($removeLegoVals['legoId'])) {
+			// echo 'Lego does not Exists';
+			throw new InvalidInputException('legonotexists');
+		}
+
+		// check if lego is in list
+		if (!$this->listEditorClass->checkLegoInLegoList($removeLegoVals['listId'], $removeLegoVals['legoId'])) {
+			// echo 'Lego not in list';
+			throw new InvalidInputException('legonotinlist');
+		}
+
 		$this->listEditorClass->deleteLegoFromList($removeLegoVals);
 	}
 
-	// post data validation methods
+	// check methods
+	protected function parentChecks(array $legoListVals): array
+	{
+		// use existing checks in LegoListContrClass to validate remaining data
+		parent::setLegoListVals(array(
+			'listName'=>$legoListVals['listName'],
+			'isPublic'=>$legoListVals['isPublic'],
+			'uid'=>$legoListVals['uid']
+		));
+		parent::legoListErrorChecks();
+
+		return parent::getLegoListVals();
+	}
 
 
 
@@ -97,11 +197,16 @@ class ListEditorContrClass {
 		$this->listEditorViewClass->echoListDataForm($this->legoListVals);
 	}
 
-	public function viewAddLegoToListForm($legoId = null): void {
-		$this->listEditorViewClass->echoAddLegoToListForm($legoId);
+	public function viewAddLegoToListForm(): void {
+		$this->listEditorViewClass->echoAddLegoToListForm();
 	}
 
 	public function viewLegosInList(): void {
 		$this->listEditorViewClass->echoLegosInList($this->legoListLegos);
+	}
+
+	public function viewLegoDB(): void {
+		$legos = $this->legoClassS->getLegos();
+		$this->listEditorViewClass->echoLegoDB($legos);
 	}
 }
